requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
google-auth>=2.23.0
requests>=2.31.0
pydantic>=2.0.0
